pyproject.toml
[project]
name = "scrape-upwork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lxml>=5.3.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.27.1",
    "undetected-chromedriver>=3.5.5",
]
